Add AstroGateway and AstroResponse classes, update tests and dependencies
Added AstroGateway class that communicates with the open-notify.org API to fetch astronaut details. To hold the data fetched, added an AstroResponse class. This setup allows us to query astronaut information and use it in our application.

A corresponding unit test, AstroGatewayTest, was also added to verify the communication with the API and the correct transformation of the data into AstroResponse objects.

To ensure JSON responses can be deserialized, we added Gson library as a dependency in build.gradle. The libraries section in libs.versions.toml was updated accordingly.