Update dependencies and modify code imports
Upgraded Spring Boot from 2.7.6 to 3.2.4 and jqwik from 1.8.3 to 1.8.4. Also, javax.persistence imports in Officer.java and JpaOfficerDAO.java have been replaced with jakarta.persistence. Minor changes have been implemented in the structure of the officers table in the schema.sql file and the test task configuration in the build.gradle.